#2225 :truck: move test util files in ergebnismeldung/common
