+++ wls-gui-wahllokalsystem/stories/components/ergebnismeldung/MBW/stapelBC/BaseTableWahlvorschlagKandidatenStimmenErfassen.stories.ts
@@ -2,7 +2,7 @@ import type { ErgebnisAndKandidat } from "@/types/ergebnisermittlung/ErgebnisAnd
 import type { Meta, StoryObj } from "@storybook/vue3";
 
 import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFactory.ts";
-import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
+import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/common/ergebnisseTestDataFactory.ts";
 import { useWahlvorschlaegeTestDataFactory } from "@tests/utils/wahlvorschlaege/WahlvorschlaegeTestDataFactory.ts";
 
 import BaseTableWahlvorschlagKandidatenStimmenErfassen from "@/components/ergebnismeldung/MBW/stapelBC/BaseTableWahlvorschlagKandidatenStimmenErfassen.vue";
