+++ wls-gui-wahllokalsystem/stories/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCSummaryCard.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/vue3";
 
-import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
+import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/common/ergebnisseTestDataFactory.ts";
 import { useWahlvorschlaegeTestDataFactory } from "@tests/utils/wahlvorschlaege/WahlvorschlaegeTestDataFactory.ts";
 
 import TheOBWStapelCSummaryCard from "@/components/ergebnismeldung/OBW/stapelC/TheOBWStapelCSummaryCard.vue";
