+++ wls-gui-wahllokalsystem/tests/components/ergebnismeldung/MBW/TheMBWUngueltigeStimmenAnzeigenCard.spec.ts
@@ -1,6 +1,6 @@
 import type { VueWrapper } from "@vue/test-utils";
 
-import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
+import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/common/ergebnisseTestDataFactory.ts";
 import { enableAutoUnmount, flushPromises, mount } from "@vue/test-utils";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { defineComponent, h, KeepAlive } from "vue";
