+++ wls-gui-wahllokalsystem/tests/composables/ergebnismeldung/common/ergebnisService.spec.ts
@@ -1,5 +1,5 @@
 import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFactory.ts";
-import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/ergebnisseTestDataFactory.ts";
+import { useErgebnisseTestDataFactory } from "@tests/utils/ergebnismeldung/common/ergebnisseTestDataFactory.ts";
 import { afterAll, afterEach, describe, expect, it, vi } from "vitest";
 
 import {
