+++ wls-gui-wahllokalsystem/tests/composables/ergebnismeldung/common/statusMapper.spec.ts
@@ -1,4 +1,4 @@
-import { useStatusTestDataFactory } from "@tests/utils/ergebnismeldung/statusTestDataFactory.ts";
+import { useStatusTestDataFactory } from "@tests/utils/ergebnismeldung/common/statusTestDataFactory.ts";
 import { describe, expect, it } from "vitest";
 
 import { MeldungDTOValidierungsstatusEnum } from "@/api/wls-clients/generated-ergebnismeldung-api";
