+++ wls-gui-wahllokalsystem/tests/stores/statusStore.spec.ts
@@ -1,5 +1,5 @@
 import { createTestingPinia } from "@pinia/testing";
-import { useStatusTestDataFactory } from "@tests/utils/ergebnismeldung/statusTestDataFactory.ts";
+import { useStatusTestDataFactory } from "@tests/utils/ergebnismeldung/common/statusTestDataFactory.ts";
 import { flushPromises } from "@vue/test-utils";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
