+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/common/StimmzettelumschlaegeTestDataFactory.ts
