+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/common/aWerteTestDataFactory.ts
