+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/common/commonErgebnismeldungTestDataFactory.ts
