+++ wls-gui-wahllokalsystem/tests/utils/ergebnismeldung/common/wahlscheineTestDataFactory.ts
@@ -4,7 +4,7 @@ import type { Builder } from "@tests/utils/Builder.ts";
 
 import { proxyBuilder } from "@tests/utils/Builder.ts";
 import { useCommonTestDataFactory } from "@tests/utils/common/CommonTestDataFactory.ts";
-import { useCommonErgebnismeldungTestDataFactory } from "@tests/utils/ergebnismeldung/commonErgebnismeldungTestDataFactory.ts";
+import { useCommonErgebnismeldungTestDataFactory } from "@tests/utils/ergebnismeldung/common/commonErgebnismeldungTestDataFactory.ts";
 
 const { generateRandomNumber } = useCommonTestDataFactory();
 const { createBezirkUndWahlID, createBezirkUndWahlIDDTO } =
